Add user rights checking, and limit the displayed log size
This adds user-rights checking to the three MissedPages activities,
using existing core rights names. We might want to change this,
but this seems likely to be okay to start with.

A limit of 100 is also added for the displayed missed pages table
so that the page doesn't grow too big. In normal usage this is
all that's required, because the top hit counts are the most
important. This will be revisited when we implement log entry
paging.

Change-Id: I7f7bf82e1a71108e482edc435823ec44d8538ce9